+++ examples/wCCD/src/utils.ts
@@ -1,11 +1,8 @@
 /* eslint-disable no-console */
 import { createContext } from 'react';
 import { detectConcordiumProvider } from '@concordium/browser-wallet-api-helpers';
-import { AccountTransactionType, GtuAmount } from '@concordium/web-sdk';
-
-export const CONTRACT_NAME_PROXY = 'CIS2-wCCD-Proxy';
-export const CONTRACT_NAME_IMPLEMENTATION = 'CIS2-wCCD';
-export const CONTRACT_NAME_STATE = 'CIS2-wCCD-State';
+import { AccountTransactionType, GtuAmount, toBuffer } from '@concordium/web-sdk';
+import { CONTRACT_NAME_PROXY } from './constants';
 
 /**
  * Action for wrapping some CCD to WCCD in the WCCD smart contract instance
@@ -39,6 +36,7 @@ export const wrap = (
                         },
                         receiveName: `${CONTRACT_NAME_PROXY}.wrap`,
                         maxContractExecutionEnergy: 30000n,
+                        parameter: toBuffer(''),
                     },
                     {
                         data: '',
@@ -94,6 +92,7 @@ export const unwrap = (
                         },
                         receiveName: `${CONTRACT_NAME_PROXY}.unwrap`,
                         maxContractExecutionEnergy: 30000n,
+                        parameter: toBuffer(''),
                     },
                     {
                         amount: amount.toString(),